split up normalize and validate tests
